Re-trigger highlights animation when highlights change
Each highlight in the data returned from the web-service contains a
timestamp when the highlight was generated. This timestamp is the
same for all the highlights, so use the first such timestamp value
as the key attribute on the motion component.

This means the animation will replay each time the highlights
change (the timestamp changed), but will not replay for example
when clicking off and back to the browser tab.