Do not hardcode server sent event end-points (as much)
Local running on port 3000 breaks with server sent events, it does
not proxy the requests to port 8080 (as per package.json).

This is why the URL was hard-coded to localhost:8080.

This means the application could only run on localhost, not from
some other device on the network.

Consequently, dynamically construct the URL using the various
components from "window.location".

This is still not ideal, since the port remains hard-coded, but is
good enough for now.