Fix back-button navigation on screens with sub-screens
Initially navigating to these screens by default will immediately
issue an automatic navigation to the configured sub-screen, this
leaves an errant item in the browser history stack.

The problem was that it was not possible to use the back button to
navigate back from those pages.

The solution is to use the "replace" option on the Navigate tag
used to do the automatic navigation.